+++ test/spec/Feature/OpenApi/DisabledOpenApiSpec.hs
@@ -3,8 +3,9 @@ module Feature.OpenApi.DisabledOpenApiSpec where
 import Network.HTTP.Types
 import Network.Wai        (Application)
 
-import Test.Hspec     hiding (pendingWith)
+import Test.Hspec
 import Test.Hspec.Wai
+import Test.Hspec.Wai.JSON
 
 import Protolude
 
@@ -13,4 +14,7 @@ spec =
   describe "Disabled OpenApi" $ do
     it "responds with 404" $
       request methodGet "/"
-        [("Accept","application/openapi+json")] "" `shouldRespondWith` 404
+        [("Accept","application/openapi+json")] ""
+        `shouldRespondWith`
+        [json| {"code":"PGRST126","details":null,"hint":null,"message":"Root endpoint metadata is disabled"} |]
+        { matchStatus = 404 }